Removing private packs now instant.
Upon removing a private pack, it now deletes the cached items and xml
file.
Store xml of the modpack on load.
Allow removal of modpacks based on xml.
Load modpacks.xml ONLY on first load. (Was loading anytime a private
pack was added as well).
Close stream on download to file.