Corrects a crash on Windows when JDK 1.7.0 is installed
As per the forum thread found below:

http://forum.feed-the-beast.com/threads/ftb-launcher-crash.35990/

This crash is due to incorrect handling of Java versions with a nonexistent build number.  In this case, the build number is omitted, and an out-of-bounds exception occurs.  This fixes that problem, and also hardens nearby codepaths in the event that they are subject to the same vulnerability.